requirements.txt
streamlit>=1.30.0
Pillow>=10.0.0
google-genai>=1.0.0
python-dotenv>=1.0.0
edge-tts>=6.1.9
pyttsx3>=2.90
thefuzz>=0.22.1
python-Levenshtein>=0.25.0
ngrok>=1.4.0
tavily-python
SpeechRecognition
pydub
streamlit-mic-recorder
omegaconf
hydra-core
fastapi
uvicorn
